Move rename user service to services package
