+++ h/admin/__init__.py
@@ -2,6 +2,4 @@
 
 
 def includeme(config):
-    config.register_service_factory('.services.user.rename_user_factory', name='rename_user')
-
     config.include('.views')
