+++ h/admin/views/users.py
@@ -9,7 +9,7 @@ from pyramid.view import view_config
 
 from h import models
 from h.accounts.events import ActivationEvent
-from h.admin.services.user import UserRenameError
+from h.services.rename_user import UserRenameError
 from h.tasks.admin import rename_user
 from memex import storage
 from h.i18n import TranslationString as _
