+++ h/services/__init__.py
@@ -9,6 +9,7 @@ def includeme(config):
     config.register_service_factory('.group.groups_factory', name='group')
     config.register_service_factory('.nipsa.nipsa_factory', name='nipsa')
     config.register_service_factory('.oauth.oauth_service_factory', name='oauth')
+    config.register_service_factory('.rename_user.rename_user_factory', name='rename_user')
     config.register_service_factory('.settings.settings_factory', name='settings')
     config.register_service_factory('.user.user_service_factory', name='user')
     config.register_service_factory('.user_signup.user_signup_service_factory', name='user_signup')
